+++ api-augment/base.ts
@@ -2,5 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // for the API, we decorate not only the endpoints, but all types
-import 'https://deno.land/x/polkadot@0.2.16/rpc-augment/mod.ts';
-import 'https://deno.land/x/polkadot@0.2.16/types-augment/mod.ts';
+import 'https://deno.land/x/polkadot/rpc-augment/mod.ts';
+import 'https://deno.land/x/polkadot/types-augment/mod.ts';