+++ rpc-provider/substrate-connect/index.ts
@@ -1,13 +1,13 @@
 // Copyright 2017-2022 @polkadot/rpc-provider authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { Config as ScConfig } from 'https://esm.sh/@substrate/connect@0.7.16';
+import type { Config as ScConfig } from 'https://esm.sh/@substrate/connect@0.7.17';
 import type { JsonRpcResponse, ProviderInterface, ProviderInterfaceCallback, ProviderInterfaceEmitCb, ProviderInterfaceEmitted } from '../types.ts';
 
-import { Chain, createScClient, ScClient, WellKnownChain } from 'https://esm.sh/@substrate/connect@0.7.16';
+import { Chain, createScClient, ScClient, WellKnownChain } from 'https://esm.sh/@substrate/connect@0.7.17';
 import EventEmitter from 'https://esm.sh/eventemitter3@4.0.7';
 
-import { isError, logger, objectSpread } from 'https://deno.land/x/polkadot@0.2.16/util/mod.ts';
+import { isError, logger, objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { RpcCoder } from '../coder/index.ts';
 import { healthChecker } from './Health.ts';