+++ types/detectOther.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2022 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { packageInfo as codecInfo } from 'https://deno.land/x/polkadot@0.2.16/types-codec/packageInfo.ts';
-import { packageInfo as createInfo } from 'https://deno.land/x/polkadot@0.2.16/types-create/packageInfo.ts';
+import { packageInfo as codecInfo } from 'https://deno.land/x/polkadot/types-codec/packageInfo.ts';
+import { packageInfo as createInfo } from 'https://deno.land/x/polkadot/types-create/packageInfo.ts';
 
 export default [codecInfo, createInfo];