+++ types/interfaces/metadata/definitions.ts
@@ -6,7 +6,7 @@
 
 import type { Definitions } from '../../types/index.ts';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.16/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { AllHashers } from './hashers.ts';
 import { runtime } from './runtime.ts';